@@ -15,11 +15,11 @@
         }
     ],
     "require": {
-        "php": ">=7.2 <9",
+        "php": "^8.1",
         "illuminate/console": "^5.5|^6.0|^7.0|^8.0|^9.0|^10.0|^11.0|^12.0",
         "illuminate/contracts": "^5.5|^6.0|^7.0|^8.0|^9.0|^10.0|^11.0|^12.0",
         "illuminate/support": "^5.5|^6.0|^7.0|^8.0|^9.0|^10.0|^11.0|^12.0",
-        "inspector-apm/inspector-php": "^3.8.0"
+        "inspector-apm/inspector-php": "3.12.3"
     },
     "autoload": {
         "psr-4": {
