@@ -17,7 +17,7 @@ class EmailServiceProvider extends ServiceProvider
      *
      * @var Segment[]
      */
-    protected $segments = [];
+    protected array $segments = [];
 
     /**
      * Booting of services.
