@@ -16,7 +16,7 @@ class GateServiceProvider extends ServiceProvider
     /**
      * @var Segment[]
      */
-    protected $segments = [];
+    protected array $segments = [];
 
     /**
      * Booting of services.
